feat: add cart items button functionality
